Move rustfix flag into its own module
