+++ src/mode.rs
@@ -2,23 +2,6 @@ use super::Error;
 use std::fmt::Display;
 use std::process::ExitStatus;
 
-/// When to run rustfix on tests
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum RustfixMode {
-    /// Do not run rustfix on the test
-    Disabled,
-    /// Apply only `MachineApplicable` suggestions emitted by the test
-    MachineApplicable,
-    /// Apply all suggestions emitted by the test
-    Everything,
-}
-
-impl RustfixMode {
-    pub(crate) fn enabled(self) -> bool {
-        self != RustfixMode::Disabled
-    }
-}
-
 #[derive(Copy, Clone, Debug)]
 /// Decides what is expected of each test's exit status.
 pub enum Mode {
