+++ src/per_test_config.rs
@@ -3,13 +3,12 @@
 //! in the files. These comments still overwrite the defaults, although
 //! some boolean settings have no way to disable them.
 
-use std::collections::HashSet;
 use std::ffi::OsString;
 use std::num::NonZeroUsize;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Output};
 
-use spanned::{Span, Spanned};
+use spanned::Spanned;
 
 use crate::custom_flags::Flag;
 use crate::dependencies::{Build, BuildManager};
@@ -20,7 +19,6 @@ use crate::rustc_stderr::Message;
 use crate::test_result::{Errored, TestOk, TestResult};
 use crate::{
     core::strip_path_prefix, rustc_stderr, Config, Error, Errors, Mode, OutputConflictHandling,
-    RustfixMode,
 };
 
 /// All information needed to run a single test
@@ -461,153 +459,7 @@ impl TestConfig<'_> {
         Ok(TestOk::Ok)
     }
 
-    pub(crate) fn run_rustfix(
-        &self,
-        output: Output,
-        global_rustfix: RustfixMode,
-    ) -> Result<bool, Errored> {
-        let no_run_rustfix = self.find_one_custom("no-rustfix")?;
-
-        let fixed_code = (no_run_rustfix.is_none() && global_rustfix.enabled())
-            .then_some(())
-            .and_then(|()| {
-                let suggestions = std::str::from_utf8(&output.stderr)
-                    .unwrap()
-                    .lines()
-                    .flat_map(|line| {
-                        if !line.starts_with('{') {
-                            return vec![];
-                        }
-                        rustfix::get_suggestions_from_json(
-                            line,
-                            &HashSet::new(),
-                            if global_rustfix == RustfixMode::Everything {
-                                rustfix::Filter::Everything
-                            } else {
-                                rustfix::Filter::MachineApplicableOnly
-                            },
-                        )
-                        .unwrap_or_else(|err| {
-                            panic!("could not deserialize diagnostics json for rustfix {err}:{line}")
-                        })
-                    })
-                    .collect::<Vec<_>>();
-                if suggestions.is_empty() {
-                    None
-                } else {
-                    let path_str = self.path.display().to_string();
-                    for sugg in &suggestions {
-                        for snip in &sugg.snippets {
-                            if snip.file_name != path_str {
-                                return Some(Err(anyhow::anyhow!("cannot apply suggestions for `{}` since main file is `{path_str}`. Please use `//@no-rustfix` to disable rustfix", snip.file_name)));
-                            }
-                        }
-                    }
-                    Some(rustfix::apply_suggestions(
-                        &std::fs::read_to_string(self.path).unwrap(),
-                        &suggestions,
-                    ))
-                }
-            })
-            .transpose()
-            .map_err(|err| Errored {
-                command: Command::new(format!("rustfix {}", self.path.display())),
-                errors: vec![Error::Rustfix(err)],
-                stderr: output.stderr,
-                stdout: output.stdout,
-            })?;
-
-        let rustfix_comments = Comments {
-            revisions: None,
-            revisioned: std::iter::once((
-                vec![],
-                Revisioned {
-                    span: Span::default(),
-                    ignore: vec![],
-                    only: vec![],
-                    stderr_per_bitwidth: false,
-                    compile_flags: self.collect(|r| r.compile_flags.iter().cloned()),
-                    env_vars: self.collect(|r| r.env_vars.iter().cloned()),
-                    normalize_stderr: vec![],
-                    normalize_stdout: vec![],
-                    error_in_other_files: vec![],
-                    error_matches: vec![],
-                    require_annotations_for_level: Default::default(),
-                    aux_builds: self.collect(|r| r.aux_builds.iter().cloned()),
-                    mode: OptWithLine::new(Mode::Pass, Span::default()),
-                    diagnostic_code_prefix: OptWithLine::new(String::new(), Span::default()),
-                    custom: self
-                        .comments
-                        .for_revision(self.revision)
-                        .flat_map(|r| r.custom.clone())
-                        .collect(),
-                },
-            ))
-            .collect(),
-        };
-        let config = TestConfig {
-            config: self.config.clone(),
-            revision: self.revision,
-            comments: &rustfix_comments,
-            path: self.path,
-        };
-
-        let run = fixed_code.is_some();
-        let mut errors = vec![];
-        let rustfix_path = config.check_output(
-            // Always check for `.fixed` files, even if there were reasons not to run rustfix.
-            // We don't want to leave around stray `.fixed` files
-            fixed_code.unwrap_or_default().as_bytes(),
-            &mut errors,
-            "fixed",
-        );
-        // picking the crate name from the file name is problematic when `.revision_name` is inserted,
-        // so we compute it here before replacing the path.
-        let crate_name = config
-            .path
-            .file_stem()
-            .unwrap()
-            .to_str()
-            .unwrap()
-            .replace('-', "_");
-        let config = TestConfig {
-            config: config.config,
-            revision: config.revision,
-            comments: &rustfix_comments,
-            path: &rustfix_path,
-        };
-        if !errors.is_empty() {
-            return Err(Errored {
-                command: Command::new(format!("checking {}", config.path.display())),
-                errors,
-                stderr: vec![],
-                stdout: vec![],
-            });
-        }
-
-        if !run {
-            return Ok(false);
-        }
-
-        let mut cmd = config.build_command()?;
-        cmd.arg("--crate-name").arg(crate_name);
-        let output = cmd.output().unwrap();
-        if output.status.success() {
-            Ok(true)
-        } else {
-            Err(Errored {
-                command: cmd,
-                errors: vec![Error::Command {
-                    kind: "rustfix".into(),
-                    status: output.status,
-                }],
-                stderr: rustc_stderr::process(&rustfix_path, &output.stderr).rendered,
-                stdout: output.stdout,
-            })
-        }
-    }
-
-    fn find_one_custom(&self, arg: &str) -> Result<OptWithLine<&dyn Flag>, Errored> {
+    pub(crate) fn find_one_custom(&self, arg: &str) -> Result<OptWithLine<&dyn Flag>, Errored> {
         self.find_one(arg, |r| r.custom.get(arg).map(|s| s.as_ref()).into())
     }
 }
